% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_qc.R
\name{allele.qc}
\alias{allele.qc}
\title{Match alleles between summary statistics and SNP information.}
\usage{
allele.qc(a1, a2, ref1, ref2)
}
\arguments{
\item{a1}{Vector of first alleles from the dataset.}

\item{a2}{Vector of alternative alleles from the dataset.}

\item{ref1}{Vector of first reference alleles.}

\item{ref2}{Vector of alternative reference alleles.}
}
\value{
A list with two elements:
\itemize{
\item \code{keep}: Logical vector indicating if the allele pair should be kept.
\item \code{flip}: Logical vector indicating if the allele pair should be flipped to match the reference.
}
}
\description{
This function performs allele flip by comparing input alleles with reference alleles.
It determines which alleles should be kept based on matching or complementarity with the reference.
The function also determines if alleles need to be flipped to match the reference.
}
