% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{create.solve_asdp_M}
\alias{create.solve_asdp_M}
\title{Approximate Solution to Semidefinite Programming for Knockoff Variable Creation}
\usage{
create.solve_asdp_M(
  Sigma,
  M = 1,
  max.size = 500,
  gaptol = 1e-06,
  maxit = 1000,
  verbose = FALSE
)
}
\description{
Approximate Solution to Semidefinite Programming for Knockoff Variable Creation
}
