% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{divide.sdp}
\alias{divide.sdp}
\title{Divide Covariance Matrix for Semidefinite Programming}
\usage{
divide.sdp(Sigma, max.size)
}
\description{
Divide Covariance Matrix for Semidefinite Programming
}
