% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.fit}
\alias{GhostKnockoff.fit}
\title{Fit Ghost Knockoff Model}
\usage{
GhostKnockoff.fit(
  Zscore_0,
  n.study,
  fit.prelim,
  gamma = 1,
  weight.study = NULL
)
}
\arguments{
\item{Zscore_0}{Matrix of Z-scores for the original variables.}

\item{n.study}{Vector of the number of studies for each variable.}

\item{fit.prelim}{List of preliminary computations from \code{GhostKnockoff.prelim}.}

\item{gamma}{Adjustment parameter for the knockoff generation, default is 1.}

\item{weight.study}{Vector of weights for each study; if NULL, weights are computed as the square root
of the number of studies divided by the square root of the sum of all studies.}
}
\value{
A list containing the weighted Z-scores for the original variables (\code{GK.Zscore_0}),
the knockoff variables (\code{GK.Zscore_k}), and the test statistics for the original (\code{T_0})
and knockoff variables (\code{T_k}).
}
\description{
Fit Ghost Knockoff Model
}
