% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.GetCorStudy}
\alias{GhostKnockoff.GetCorStudy}
\title{Calculate Study-Specific Correlation Matrix}
\usage{
GhostKnockoff.GetCorStudy(Zscore_0, fit.prelim)
}
\description{
Calculate Study-Specific Correlation Matrix
}
