% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.prelim.Meta}
\alias{GhostKnockoff.prelim.Meta}
\title{Meta Analysis for Preliminary Ghost Knockoff Computations}
\usage{
GhostKnockoff.prelim.Meta(cor.study, n.study)
}
\description{
Meta Analysis for Preliminary Ghost Knockoff Computations
}
