% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.prelim}
\alias{GhostKnockoff.prelim}
\title{Preliminary Computations for Ghost Knockoff Variables}
\usage{
GhostKnockoff.prelim(cor.G, M = 5, method = "asdp", max.size = 500)
}
\arguments{
\item{cor.G}{A covariance matrix representing the relationships between genes.}

\item{M}{The number of knockoff variables to generate, default is 5.}

\item{method}{The method used for solving the optimization problem, either 'sdp' or 'asdp', default is 'asdp'.}

\item{max.size}{The maximum size for the semidefinite programming problem, default is 500.}
}
\value{
A list containing the P.each matrix, the V.left matrix (left null space), the Normal_50Studies matrix,
and permute.index, which are used in the knockoff variable generation process.
}
\description{
This function performs preliminary computations needed to generate Ghost Knockoff variables.
It takes a covariance matrix of genes, the number of knockoff variables to generate,
the method for solving the semidefinite programming problem, and an optional size limit
for the problem. It returns matrices necessary for the subsequent steps in the Ghost
Knockoff generation process.
}
