% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS_GKF.R
\name{twas}
\alias{twas}
\title{Perform Transcriptome-Wide Association Study (TWAS)}
\usage{
twas(genomic_data, phenotype_data)
}
\arguments{
\item{genomic_data}{Matrix of genomic data where rows are samples and columns are genetic features.}

\item{phenotype_data}{Numeric vector of phenotype data where each element corresponds to a sample.}
}
\value{
A list with two elements:
\describe{
\item{\code{wgt.matrix}}{Matrix of effect weights from the best model, zeroed if performance criteria are not met.}
\item{\code{pred.wgt}}{Matrix of predicted values based on the best model, zeroed if performance criteria are not met.}
}
}
\description{
This function conducts a TWAS using genomic data to assess its association with a phenotype.
It applies elastic net regularization to build a predictive model.
Cross-validation is used to evaluate the model, and the function outputs the weights and
predictions for the best model based on performance metrics.
}
