% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS_GKF.R
\name{TWAS_GKF}
\alias{TWAS_GKF}
\title{Transcriptome-Wide Association Study with GhostKnockoff Filter}
\usage{
TWAS_GKF(weight, LD, sumstats, gene_matrix, n.study, M = 5, FDR_level = 0.05)
}
\arguments{
\item{weight}{A numeric vector of SNP weights for predicting gene expression.}

\item{LD}{The linkage disequilibrium matrix for the SNPs.}

\item{sumstats}{A data frame or matrix containing summary statistics from GWAS, with the columns "beta" and "se", including beta coefficients and standard errors.}

\item{gene_matrix}{A matrix of gene expression data.}

\item{n.study}{A vactor with the number of studies or samples.}

\item{M}{The number of knockoff variables to be generated, default is 5.}

\item{FDR_level}{The desired false discovery rate level, default is 0.05.}
}
\value{
A list containing:
\itemize{
\item \code{var}: Variance of the gene expression regulated by the SNPs.
\item \code{est}: Effect size estimate.
\item \code{z}: Z-score for the association.
\item \code{qvalue}: qß-values from the knockoff filter.
\item \code{gene_select}: Indices of the selected genes passing the FDR threshold.
}
}
\description{
This function performs a TWAS using the GhostKnockoff Filter (GKF) to control the false discovery rate.
It calculates the variance and effect size estimate of gene expression based on genetic variants,
and then determines the significance of each gene's association with the phenotype using z-scores and knockoff statistics.
}
